mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cameo-mcp-server"
version = "2.3.0"
description = "MCP server bridging AI assistants to CATIA Magic / Cameo Systems Modeler"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Cole Lyons" },
]
keywords = ["mcp", "cameo", "catia-magic", "sysml", "uml", "mbse"]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "Pillow>=10.0.0",
    "python-pptx>=1.0.2",
]

[project.urls]
Homepage = "https://github.com/ajhcs/cameo-mcp-bridge"
Repository = "https://github.com/ajhcs/cameo-mcp-bridge"
Issues = "https://github.com/ajhcs/cameo-mcp-bridge/issues"

[project.scripts]
cameo-mcp = "cameo_mcp.server:main"

[tool.setuptools]
include-package-data = false

[tool.setuptools.packages.find]
where = ["."]
include = ["cameo_mcp*"]
